furnace module monthlySummary component
modify the value passed into the daily summaryItem component.  Problem fixed: when iterating a range, the starting value is 1 instead of '0' (as in looping through array index), so the starting date had been started on the 2nd instead of the 1st of the month.  Therefore the shipments on the first of the month were not rendered, and also the records for the first of the 'next' month was accidently rendered as a result